+++ if-file-exists.php
@@ -49,6 +49,8 @@ if ( ! function_exists( 'c2c_if_file_exists' ) ) :
  *   %file_directory% : the directory of the file, i.e. "/usr/local/www/yoursite/wp-content/uploads/"
  *   %file_extension% : the extension of the file, i.e. "zip"
  *   %file_name%      : the name of the file, i.e. "pictures.zip"
+ *   %file_size%      : the size of the file, in number of bytes to the largest unit the bytes will fit into, rounded to 2 decimal places, i.e. "1.24 MB"
+ *   $file_size_byees%: the size of the file, in bytes
  *   %file_url%       : the URL of the file, i.e. "http://yoursite.com/wp-content/uploads/pictures.zip";
  *   %file_path%      : the filesystem path to the file, i.e. "/usr/local/www/yoursite/wp-content/uploads/pictures.zip"
  *
@@ -123,6 +125,8 @@ function c2c_if_file_exists( $filename, $format = '', $echo = true, $dir = '', $
 				'%file_extension%' => isset( $pathparts['extension'] ) ? $pathparts['extension'] : '',
 				'%file_name%'      => $pathparts['basename'],
 				'%file_path%'      => $full_path,
+				'%file_size%'      => str_replace( '.00', '', size_format( wp_filesize( $full_path ), 2 ) ),
+				'%file_size_bytes%'=> wp_filesize( $full_path ),
 				'%file_url%'       => site_url() . '/' . $dir . '/' . $filename
 			);